refactor(core): split search/ranking into submodules
Split large ranking.rs file into module directory structure:
- mod.rs: Public API and exports
- scoring.rs: Ranking score calculations
- wilson.rs: Wilson score interval implementation
- tests.rs: Unit tests

File size compliance: Each file now <500 LOC.
